+++ src/ui/create_user_view.py
@@ -1,23 +1,28 @@
-from tkinter import ttk
+from tkinter import ttk, StringVar
 from repositories.user_repository import user_repository
-from services.magic_service import MagicService
+from services.magic_service import MagicService, UsernameExistsError
 
 
 class CreateUserView:
     """User interface for creating a new user."""
 
-    def __init__(self, root):
+    def __init__(self, root, show_login_view):
         """Class constructor. Creates a new view for creating a new user.
 
         Args:
             root:
                 TKinter -element, which initializes the user interface.
+            show_login_view:
+                Direct user to login view
         """
 
         self._root = root
+        self._show_login_view = show_login_view
         self._frame = None
         self._new_username_entry = None
         self._new_password_entry = None
+        self._error_variable = None
+        self._error_label = None
 
         # Center the window
         self._height = 500
@@ -41,14 +46,23 @@ class CreateUserView:
         password = self._new_password_entry.get()
         magic_service = MagicService(user_repository)
 
-        if len(username) == 0 or len(password) == 0:
-            print("Username or password is missing!")
+        if len(username) < 3:
+            self._show_error("Username must be at least 3 characters long")
+            return
+        if len(password) < 12:
+            self._show_error("Password must be at least 12 characters long")
             return
 
         try:
             magic_service.create_user(username, password)
-        except:
-            print("Error in creating user!")
+            # TBA: "Account created successfully -alert"
+            self._show_login_view()
+        except UsernameExistsError as e:
+            self._show_error(str(e))
+    
+    def _show_error(self, message):
+        self._error_variable.set(message)
+        self._error_label.pack()
 
     def _initialize_new_username(self):
         new_username_label = ttk.Label(master=self._frame, text="Username")
@@ -66,6 +80,7 @@ class CreateUserView:
 
     def _initialize(self):
         self._frame = ttk.Frame(master=self._root)
+        self._error_variable = StringVar(self._frame)
 
         title_label = ttk.Label(
             master=self._frame,
@@ -82,3 +97,12 @@ class CreateUserView:
             command=self._create_user_handler,
             width=20)
         create_user_button.pack(pady=(15, 0))
+
+        self._error_label = ttk.Label(
+            master=self._frame,
+            textvariable=self._error_variable,
+            foreground="red"
+        )
+        self._error_label.pack(pady=(5, 0))
+
+        self._error_label.pack_forget()